Fix Invoke-Kerberoast with etype 17 or 18
AES-encrypted Kerberos service tickets (etype 17 or etype 18) use a
different length checksum. This can be seen easiest in the source code
of impacket:

<https://github.com/fortra/impacket/blob/32178de69075ba51d386a2973975e30533c2edd3/examples/GetUserSPNs.py#L191..L229>

It is 16 bytes for RC4 or DES tickets and 12 bytes for AES tickets.
Since the code is parsing hexascii encoded binary data, the values need
to be doubled. The syntax of the hash is such that a dollar sign
separates the checksum from the rest of the data.

This patch inserts the dollar sign at the correct position for etypes 17
and 18.